report by product without status

+++ account-service/src/main/java/com/ironhack/accountservice/controller/dtos/report/OpportunityBySalesRepDTO.java
@@ -2,14 +2,14 @@ package com.ironhack.accountservice.controller.dtos.report;
 
 public class OpportunityBySalesRepDTO {
     private Integer salesRepId;
-    private Integer count;
+    private Long count;
 
     public OpportunityBySalesRepDTO() {
     }
 
-    public OpportunityBySalesRepDTO(Integer salesRepId, Integer count) {
+    public OpportunityBySalesRepDTO(Integer salesRepId, Long count) {
         this.salesRepId = salesRepId;
-        this.count = count;
+        this.count = count.longValue();
     }
 
     public Integer getSalesRepId() {
@@ -20,11 +20,11 @@ public class OpportunityBySalesRepDTO {
         this.salesRepId = salesRepId;
     }
 
-    public Integer getCount() {
+    public Long getCount() {
         return count;
     }
 
-    public void setCount(Integer count) {
+    public void setCount(Long count) {
         this.count = count;
     }
 }
